+++ packages/components/grid/src/filter.ts
@@ -1,9 +1,9 @@
 import type { CSSResultGroup, TemplateResult } from 'lit';
 import type { GridColumn } from './column.js';
-import type { GridFilterMode, GridFilterOption } from './filter-column.js';
 import type { ScopedElementsMap } from '@open-wc/scoped-elements';
-import type { EventEmitter } from '@sl-design-system/shared';
-import { faFilter, faFilterList, faXmark } from '@fortawesome/pro-regular-svg-icons';
+import type { DataSourceFilterFunction, EventEmitter } from '@sl-design-system/shared';
+import { faFilter, faXmark } from '@fortawesome/pro-regular-svg-icons';
+import { faFilter as faFilterSolid } from '@fortawesome/pro-solid-svg-icons';
 import { localized, msg } from '@lit/localize';
 import { ScopedElementsMixin } from '@open-wc/scoped-elements';
 import { Button } from '@sl-design-system/button';
@@ -14,20 +14,16 @@ import { Popover } from '@sl-design-system/popover';
 import { event, getNameByPath } from '@sl-design-system/shared';
 import { LitElement, html } from 'lit';
 import { property } from 'lit/decorators.js';
+import { type GridFilterMode, type GridFilterOption } from './filter-column.js';
 import styles from './filter.scss.js';
+import { GridFilterValueChangeEvent } from './events.js';
 
 export type GridFilterChange = 'added' | 'removed';
 
-Icon.registerIcon(faFilter, faFilterList, faXmark);
-
-export class GridFilterValueChangeEvent extends Event {
-  constructor(public readonly column: GridColumn, public readonly value: string | string[] | undefined) {
-    super('sl-filter-value-change', { bubbles: true, composed: true });
-  }
-}
+Icon.registerIcon(faFilter, faFilterSolid, faXmark);
 
 @localized()
-export class GridFilter extends ScopedElementsMixin(LitElement) {
+export class GridFilter<T> extends ScopedElementsMixin(LitElement) {
   /** @private */
   static get scopedElements(): ScopedElementsMap {
     return {
@@ -53,15 +49,25 @@ export class GridFilter extends ScopedElementsMixin(LitElement) {
   @property({ type: Boolean, reflect: true }) active = false;
 
   /** The grid column. */
-  @property({ attribute: false }) column!: GridColumn;
+  @property({ attribute: false }) column!: GridColumn<T>;
 
+  /** The custom filter */
+  @property({ attribute: false }) filter?: DataSourceFilterFunction<T>;
+
+  /** Emits when the filter has been added or removed. */
   @event() filterChange!: EventEmitter<GridFilterChange>;
 
-  @event() filterValueChange!: EventEmitter<GridFilterValueChangeEvent>;
+  /** Emits when the value of the this filter has changed. */
+  @event() filterValueChange!: EventEmitter<GridFilterValueChangeEvent<T>>;
 
+  /** The mode of the filter. */
   @property({ type: String }) mode?: GridFilterMode;
 
+  /** The filter options. */
   @property({ attribute: false }) options?: GridFilterOption[];
+
+  /** The path to the field to filter on. */
+  @property() path?: string;
 
   set value(value: string | string[] | undefined) {
     if (this.mode !== 'text') {
@@ -93,22 +99,22 @@ export class GridFilter extends ScopedElementsMixin(LitElement) {
 
   override render(): TemplateResult {
     return html`
-      <sl-button class="toggle" id="anchor" fill="link">
+      <sl-button @click=${this.#onClick} class="toggle" id="anchor" fill="link">
         <slot></slot>
-        <sl-icon .name=${this.active ? 'far-filter-list' : 'far-filter'}></sl-icon>
+        <sl-icon .name=${this.active ? 'fas-filter' : 'far-filter'}></sl-icon>
       </sl-button>
-      <sl-popover anchor="anchor">
+      <sl-popover anchor="anchor" position="bottom">
         <header>
           <h1 id="title">
             ${msg(`Filter by`)} <span>${this.column.header?.toString() || getNameByPath(this.column.path)}</span>
           </h1>
-          <sl-button @click=${this.#onClick} aria-label=${msg('Close')} fill="link" size="sm">
+          <sl-button @click=${this.#onHide} aria-label=${msg('Close')} fill="link" size="sm">
             <sl-icon name="far-xmark"></sl-icon>
           </sl-button>
         </header>
         ${this.mode === 'select'
           ? html`
-              <sl-checkbox-group aria-labelledby="title">
+              <sl-checkbox-group aria-labelledby="title" autofocus>
                 ${this.options?.map(
                   option =>
                     html`
@@ -125,10 +131,12 @@ export class GridFilter extends ScopedElementsMixin(LitElement) {
             `
           : html`
               <sl-text-input
+                @keydown=${this.#onKeydown}
                 @input=${this.#onInput}
                 .placeholder=${msg('Type here to filter')}
                 .value=${this.value?.toString() ?? ''}
                 aria-labelledby="title"
+                autofocus
               ></sl-text-input>
             `}
       </sl-popover>
@@ -150,7 +158,18 @@ export class GridFilter extends ScopedElementsMixin(LitElement) {
   }
 
   #onClick(): void {
+    this.renderRoot.querySelector<Popover>('sl-popover')?.togglePopover();
+  }
+
+  #onHide(): void {
     this.renderRoot.querySelector<Popover>('sl-popover')?.hidePopover();
+  }
+
+  #onKeydown(event: KeyboardEvent): void {
+    if (event.key === 'Enter') {
+      event.preventDefault();
+      this.#onClick();
+    }
   }
 
   #onInput(event: Event & { target: HTMLInputElement }): void {