+++ packages/components/grid/src/stories/filtering.stories.ts
@@ -38,10 +38,29 @@ export const Filtered: Story = {
   `
 };
 
+export const EmptyValues: Story = {
+  render: () => {
+    const items = [
+      { key: 'Foo', value: 'foo' },
+      { key: 'Spaces', value: '          ' },
+      { key: 'Blank', value: '' },
+      { key: 'Null', value: null },
+      { key: 'Undefined', value: undefined }
+    ];
+
+    return html`
+      <sl-grid .items=${items} style="width: 200px">
+        <sl-grid-column path="key"></sl-grid-column>
+        <sl-grid-filter-column path="value"></sl-grid-filter-column>
+      </sl-grid>
+    `;
+  }
+};
+
 export const OutsideGrid: Story = {
   render: (_, { loaded: { people } }) => {
     const onInput = ({ target }: Event & { target: TextInput }): void => {
-      const grid = document.querySelector('sl-grid') as Grid,
+      const grid = document.querySelector('sl-grid') as Grid<Person>,
         regex = new RegExp(target.value?.toString().trim() ?? '', 'i');
 
       grid.items = (people as Person[]).filter(({ firstName, lastName, email, profession }) => {