+++ packages/components/shared/src/controllers/index.ts
@@ -1,3 +1,4 @@
+export * from './anchor.js';
 export * from './events.js';
 export * from './focus-group.js';
 export * from './roving-tabindex.js';